+++ src/controllers/AccountController.php
@@ -34,6 +34,7 @@ class AccountController extends Controller
     public function behaviors(): array
     {
         return [
+            ...parent::behaviors(),
             'access' => [
                 'class' => AccessControl::class,
                 'rules' => [