+++ src/models/Redirect.php
@@ -34,10 +34,11 @@ class Redirect extends ActiveRecord implements TypeAttributeInterface
 
     public function behaviors(): array
     {
-        return array_merge(parent::behaviors(), [
+        return [
+            ...parent::behaviors(),
             'DateTimeBehavior' => DateTimeBehavior::class,
             'TrailBehavior' => TrailBehavior::class,
-        ]);
+        ];
     }
 
     public function rules(): array