+++ src/modules/admin/controllers/RedirectController.php
@@ -17,6 +17,7 @@ class RedirectController extends Controller
     public function behaviors(): array
     {
         return [
+            ...parent::behaviors(),
             'access' => [
                 'class' => AccessControl::class,
                 'rules' => [