+++ src/modules/admin/controllers/TrailController.php
@@ -14,6 +14,7 @@ class TrailController extends Controller
     public function behaviors(): array
     {
         return [
+            ...parent::behaviors(),
             'access' => [
                 'class' => AccessControl::class,
                 'rules' => [