+++ src/modules/admin/controllers/UserController.php
@@ -25,6 +25,7 @@ class UserController extends Controller
     public function behaviors(): array
     {
         return [
+            ...parent::behaviors(),
             'access' => [
                 'class' => AccessControl::class,
                 'rules' => [