+++ src/web/UrlManager.php
@@ -2,13 +2,14 @@
 
 namespace davidhirtz\yii2\skeleton\web;
 
+use davidhirtz\yii2\skeleton\helpers\ArrayHelper;
 use Yii;
 use yii\web\UrlRule;
 
 class UrlManager extends \yii\web\UrlManager
 {
-    public $enablePrettyUrl = true;
-    public $showScriptName = false;
+    public const EVENT_AFTER_CREATE = 'afterCreate';
+    public const EVENT_BEFORE_PARSE = 'beforeParse';
 
     /**
      * @var bool whether the language should be added to the URL via `languageParam`.
@@ -21,8 +22,8 @@ class UrlManager extends \yii\web\UrlManager
     public bool $i18nSubdomain = false;
 
     /**
-     * @var array|null containing the languages available for `i18nUrl` or `i18nSubdomain`, leave empty to use languages
-     * defined in `i18n` component.
+     * @var array|null containing the languages available for `i18nUrl` or `i18nSubdomain`, the language identifier as
+     * key and the language param as value (e.g. ['en-US' => 'en']). Defaults to languages set in the I18n component.
      */
     public ?array $languages = null;
 
@@ -33,11 +34,6 @@ class UrlManager extends \yii\web\UrlManager
     public string|false|null $defaultLanguage = null;
 
     /**
-     * @var string the name of the GET parameter that specifies the language.
-     */
-    public string $languageParam = 'language';
-
-    /**
      * @var array containing hard redirects, either as request URI => URL pairs, which generate regular 301 redirects
      * or as arrays containing the request URIs at first position, the target URL as the second and an
      * optional third containing the redirect code (defaults to 301). If dynamic redirects are needed, please take
@@ -45,13 +41,9 @@ class UrlManager extends \yii\web\UrlManager
      */
     public array $redirectMap = [];
 
-    /**
-     * Events.
-     */
-    public const EVENT_AFTER_CREATE = 'afterCreate';
-    public const EVENT_BEFORE_PARSE = 'beforeParse';
+    public $enablePrettyUrl = true;
+    public $showScriptName = false;
 
-    
     public function init(): void
     {
         if (!$this->enablePrettyUrl) {
@@ -65,6 +57,8 @@ class UrlManager extends \yii\web\UrlManager
         $this->defaultLanguage ??= Yii::$app->sourceLanguage;
 
         if ($this->languages === null) {
+            $this->languages = [];
+
             foreach (Yii::$app->getI18n()->getLanguages() as $language) {
                 $this->languages[$language] = strstr((string)$language, '-', true) ?: $language;
             }
@@ -83,6 +77,7 @@ class UrlManager extends \yii\web\UrlManager
 
     public function createUrl($params): string
     {
+        $request = Yii::$app->getRequest();
         $language = Yii::$app->language;
         $i18nUrl = $this->i18nUrl;
 
@@ -92,36 +87,30 @@ class UrlManager extends \yii\web\UrlManager
         }
 
         if ($i18nUrl || $this->i18nSubdomain) {
-            if (isset($params[$this->languageParam])) {
-                $language = $params[$this->languageParam];
-                unset($params[$this->languageParam]);
-            }
+            $language = ArrayHelper::remove($params, $request->languageParam, $language);
         }
 
         $url = parent::createUrl(array_filter($params, fn ($value): bool => !is_null($value)));
 
-        $this->trigger(static::EVENT_AFTER_CREATE, $event = Yii::$container->get(UrlManagerEvent::class, [], [
-            'url' => $url,
-            'params' => $params,
-        ]));
+        $event = $this->getAfterCreateEvent($url, $params);
+        $url = $event?->url ?? $url;
 
         if ($i18nUrl) {
             if (isset($this->languages[$language]) && $language !== $this->defaultLanguage) {
                 $position = strlen($this->showScriptName ? $this->getScriptUrl() : $this->getBaseUrl());
-                return rtrim(substr_replace((string) $event->url, '/' . $this->languages[$language], $position, 0), '/');
+                return rtrim(substr_replace($url, '/' . $this->languages[$language], $position, 0), '/');
             }
         }
 
         if ($this->i18nSubdomain && $language !== Yii::$app->language) {
-            $request = Yii::$app->getRequest();
-            $subdomain = $language == $this->defaultLanguage || !in_array($language, $this->languages) ?
-                ($request->getIsDraft() ? $request->draftSubdomain : 'www') :
-                ($request->getIsDraft() ? ($request->draftSubdomain . '.' . $language) : $language);
+            $subdomain = $language == $this->defaultLanguage || !in_array($language, $this->languages)
+                ? ($request->getIsDraft() ? $request->draftSubdomain : 'www')
+                : ($request->getIsDraft() ? ($request->draftSubdomain . '.' . $language) : $language);
 
-            return parse_url($this->getHostInfo(), PHP_URL_SCHEME) . '://' . $subdomain . $this->getI18nHostInfo() . $event->url;
+            return parse_url($this->getHostInfo(), PHP_URL_SCHEME) . '://' . $subdomain . $this->getI18nHostInfo() . $url;
         }
 
-        return $event->url;
+        return $url;
     }
 
     public function createDraftUrl(array $params): string
@@ -138,11 +127,20 @@ class UrlManager extends \yii\web\UrlManager
      */
     public function parseRequest($request): bool|array
     {
-        $pathInfo = trim($request->getPathInfo(), '/');
-        $response = Yii::$app->getResponse();
+        $this->parseRedirectMap($request, $this->redirectMap);
+        $this->setApplicationLanguage($request);
 
-        if ($this->redirectMap) {
-            foreach ($this->redirectMap as $urlset => $location) {
+        $event = $this->getBeforeParseEvent($request);
+
+        return parent::parseRequest($event?->request ?? $request);
+    }
+
+    protected function parseRedirectMap(Request $request, array $redirectMap): void
+    {
+        if ($redirectMap) {
+            $pathInfo = trim($request->getPathInfo(), '/');
+
+            foreach ($redirectMap as $urlset => $location) {
                 $statusCode = 301;
 
                 if (is_array($location)) {
@@ -163,51 +161,83 @@ class UrlManager extends \yii\web\UrlManager
                     $wildcard = strpos($url, '*');
 
                     if ($url == $pathInfo || ($wildcard && substr($url, 0, $wildcard) == substr($pathInfo, 0, $wildcard))) {
-                        $response->redirect($location, $statusCode);
+                        Yii::$app->getResponse()->redirect($location, $statusCode);
                         Yii::$app->end();
                     }
                 }
             }
         }
+    }
 
+    protected function setApplicationLanguage(Request $request): void
+    {
         if ($this->i18nUrl) {
             // Check if the pathInfo starts with a language identifier.
+            $pathInfo = trim($request->getPathInfo(), '/');
+
             if (preg_match('#^(' . implode('|', $this->languages) . ')\b(/?)#i', $pathInfo, $matches)) {
                 $request->setPathInfo(mb_substr($pathInfo, mb_strlen($matches[0], Yii::$app->charset), null, Yii::$app->charset));
                 $language = array_search($matches[1], $this->languages);
 
                 if ($language) {
                     if ($language == $this->defaultLanguage) {
-                        $response->redirect($request->getHostInfo() . '/' . $request->getPathInfo(), 301);
+                        $url = $request->getHostInfo() . '/' . $request->getPathInfo();
+                        Yii::$app->getResponse()->redirect($url, 301);
                         Yii::$app->end();
                     }
 
                     Yii::$app->language = $language;
                 }
-            } else {
-                Yii::$app->language = $this->defaultLanguage ?: $request->getPreferredLanguage(array_keys($this->languages));
+
+                return;
             }
         }
 
         if ($this->i18nSubdomain) {
-            $subdomain = explode('.', parse_url($this->getHostInfo(), PHP_URL_HOST))[$request->getIsDraft() ? 1 : 0];
-            Yii::$app->language = in_array($subdomain, $this->languages) ? $subdomain : $this->defaultLanguage;
+            $host = parse_url($this->getHostInfo(), PHP_URL_HOST);
+            $subdomain = explode('.', $host)[$request->getIsDraft() ? 1 : 0];
+
+            if (in_array($subdomain, $this->languages)) {
+                Yii::$app->language = $subdomain;
+                return;
+            }
         }
 
-        $this->trigger(static::EVENT_BEFORE_PARSE, $event = Yii::$container->get(UrlManagerEvent::class, [], [
-            'request' => $request,
-        ]));
+        if (in_array($request->getLanguage(), $this->languages)) {
+            Yii::$app->language = $request->getLanguage();
+            return;
+        }
 
-        return parent::parseRequest($event->request);
+        Yii::$app->language = $this->defaultLanguage ?: $request->getPreferredLanguage(array_keys($this->languages));
+    }
+
+    protected function getBeforeParseEvent(Request $request): ?UrlManagerEvent
+    {
+        $event = Yii::$container->get(UrlManagerEvent::class, [], ['request' => $request]);
+        $this->trigger(static::EVENT_BEFORE_PARSE, $event);
+
+        return $event;
+    }
+
+    protected function getAfterCreateEvent(string $url, array $params): ?UrlManagerEvent
+    {
+        $event = Yii::$container->get(UrlManagerEvent::class, [], [
+            'url' => $url,
+            'params' => $params,
+        ]);
+
+        $this->trigger(static::EVENT_AFTER_CREATE, $event);
+
+        return $event;
     }
 
     /**
      * Generates a list of rule parameters at given position. This can be used to validate dynamic slugs, etc.
-     * @noinspection PhpUnused
      */
     public function getImmutableRuleParams(int $position = 0): array
     {
         $params = [];
+
         foreach (Yii::$app->getUrlManager()->rules as $rule) {
             if ($rule instanceof UrlRule) {
                 $param = explode('/', $rule->name)[$position];